Extract pick_from to helpers and ask to continue if getting examples fails
